updated some css
