+++ src/components/form/form.scss
@@ -1,0 +1,5 @@
+Form {
+  border: 1px solid rgb(164, 164, 164);
+  padding: 1rem;
+  border-radius: 0.2rem;
+}
