+++ src/components/small-nav/small-nav.scss
@@ -1,0 +1,4 @@
+.small-nav {
+  width: 80%;
+  margin: 1rem auto;
+}
