+++ src/styles.scss
@@ -1,9 +1,3 @@
-// * {
-//   background-color: bisque;
-
-// }
-
-// .testing {
-//   background-color: red;
-
-// }
+.main {
+  width: 80%;
+}
